+++ plugins/ember.js
@@ -3,13 +3,15 @@
  *
  * Patches event handler callbacks and ajax callbacks.
  */
-;(function(window, Raven, Ember) {
+;(function(window) {
 'use strict';
 
+if (window.Raven) Raven.addPlugin(function EmberPlugin() {
+
+var Ember = window.Ember;
+
 // quit if Ember isn't on the page
-if (!Ember) {
-    return;
-}
+if (!Ember) return;
 
 var _oldOnError = Ember.onerror;
 Ember.onerror = function EmberOnError(error) {
@@ -26,4 +28,7 @@ Ember.RSVP.on('error', function (reason) {
     }
 });
 
-}(window, window.Raven, window.Ember));
+// End of plugin factory
+});
+
+}(typeof window !== 'undefined' ? window : this));